--- src/components/PostUploader/ImageUploader.scss
@@ -1,117 +1,0 @@
-@import "../../stylesheets/util.scss";
-
-.image-uploader {
-  display: flex;
-  justify-content: space-around;
-  flex-wrap: wrap;
-  width: 60rem;
-
-  &-preview {
-    &-wrapper {
-      position: relative;
-
-      & + & {
-        margin-left: 2rem;
-      }
-    }
-
-    &-img {
-      width: 15rem;
-      height: 15rem;
-      margin-bottom: 3rem;
-      border: 1px solid $main-color-light;
-
-      &.representative {
-        border: 3px solid $main-color-light;
-      }
-
-      &:hover {
-        cursor: pointer;
-      }
-
-      &-close-btn {
-        position: absolute;
-        right: 0.8rem;
-        top: 0.8rem;
-        width: 2.2rem;
-      }
-
-      &-select-btn {
-        position: absolute;
-        top: 0.8rem;
-        left: 0.8rem;
-        width: 2.2rem;
-      }
-    }
-  }
-
-  &-first-input {
-    &-wrapper {
-      padding: 0 1.5rem;
-      height: 4rem;
-      line-height: 3.5rem;
-      border: 1px solid $main-color;
-      border-radius: 2rem;
-      background-color: $main-color-light;
-      display: flex;
-      align-items: center;
-      justify-content: center;
-    }
-
-    &-icon {
-      max-width: 80%;
-      max-height: 80%;
-    }
-
-    &-label {
-      font-weight: bold;
-      color: white;
-      font-size: 1.5rem;
-    }
-
-    &-btn {
-      position: absolute;
-      width: 1px;
-      height: 1px;
-      padding: 0;
-      margin: -1px;
-      overflow: hidden;
-      clip: rect(0, 0, 0, 0);
-      border: 0;
-    }
-  }
-
-  &-second-input {
-    &-wrapper {
-      width: 15rem;
-      height: 15rem;
-      margin-left: 2rem;
-      border: 2px dashed $main-color-light;
-      display: flex;
-      align-items: center;
-      justify-content: center;
-    }
-
-    &-label {
-      display: inline-block;
-      padding: 1.5rem 2rem;
-      border-radius: 50%;
-      background-color: $main-color-light;
-      color: white;
-      font-size: 3rem;
-      line-height: normal;
-      cursor: pointer;
-    }
-
-    &-btn {
-      position: absolute;
-      width: 1px;
-      height: 1px;
-      padding: 0;
-      margin: -1px;
-      overflow: hidden;
-      clip: rect(0, 0, 0, 0);
-      border: 0;
-    }
-  }
-}
