+++ src/components/PostUploader/ImageUploader/PreviewImages.scss
@@ -1,0 +1,40 @@
+@import "../../../stylesheets/util.scss";
+
+.image-uploader-preview {
+  &-wrapper {
+    position: relative;
+
+    & + & {
+      margin-left: 2rem;
+    }
+  }
+
+  &-img {
+    width: 15rem;
+    height: 15rem;
+    margin-bottom: 3rem;
+    border: 1px solid $main-color-light;
+
+    &.representative {
+      border: 3px solid $main-color-light;
+    }
+
+    &:hover {
+      cursor: pointer;
+    }
+
+    &-close-btn {
+      position: absolute;
+      right: 0.8rem;
+      top: 0.8rem;
+      width: 2.2rem;
+    }
+
+    &-select-btn {
+      position: absolute;
+      top: 0.8rem;
+      left: 0.8rem;
+      width: 2.2rem;
+    }
+  }
+}
