+++ src/components/PostUploader/ImageUploader/SecondInputButton.scss
@@ -1,0 +1,35 @@
+@import "../../../stylesheets/util.scss";
+
+.image-uploader-second-input {
+  &-wrapper {
+    width: 15rem;
+    height: 15rem;
+    margin-left: 2rem;
+    border: 2px dashed $main-color-light;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+  }
+
+  &-label {
+    display: inline-block;
+    padding: 1.5rem 2rem;
+    border-radius: 50%;
+    background-color: $main-color-light;
+    color: white;
+    font-size: 3rem;
+    line-height: normal;
+    cursor: pointer;
+  }
+
+  &-btn {
+    position: absolute;
+    width: 1px;
+    height: 1px;
+    padding: 0;
+    margin: -1px;
+    overflow: hidden;
+    clip: rect(0, 0, 0, 0);
+    border: 0;
+  }
+}
